Modify crc32Table to uint
When building with options GOARCH=386 the hex values in this table that start with a hex 9-F overflow the int value.  The error message is:
"cannot use 0x9823B6E0 (untyped int constant 2552477408) as int value in array or slice literal (overflows)"
Changing the value type to 'uint' fixes the build issue.